requirements.txt
msal==1.26.0
platformdirs==4.1.0
pytz==2023.3.post1
PyYAML==6.0.1
Requests==2.31.0
